accesor functions for objects of SignalpResult S4 class, intermediate and final outputs of the signalp prediction step

accesor functions for objects of SignalpResult S4 class, intermediate and final outputs of the signalp prediction step

Accessors for SignalpResult objects

setMatfasta(theObject, mature_fasta)

# S4 method for SignalpResult
setMatfasta(theObject, mature_fasta)

# S4 method for SignalpResult
getMatfasta(theObject)

# S4 method for SignalpResult
setSPversion(theObject, sp_version)

# S4 method for SignalpResult
getSPversion(theObject)

# S4 method for SignalpResult
setSPtibble(theObject, sp_tibble)

# S4 method for SignalpResult
getSPtibble(theObject)

Arguments

theObject
SignalpResult object

mature_fasta
sequences with clipped signal peptides, AAStringSet object

sp_version
version of signalp used to generate SignalpResult object

sp_tibble
parsed signalp output in tabular format

Value

SignalpResult object

Slots

mature_fasta

fasta with mature sequences

sp_version

version of signalp used to generate this object

sp_tibble

Object of class tibble, columns:

gene_id - unique id of the sequence

Cmax - max raw cleavage site score (C-score)

Cpos - amino acid position with max C-score

Cparsed - sp2 and sp3: remove this field?

Ymax - max combied cleavage site score (Y-score)

Ypos - amino acid position with max Y-score

Smax - max signal peptide score

Spos - amino acid position with max S-score

Smean - the average S-score of the possible signal peptide (from position 1 to the position immediately before the maximal Y-score)

Prediction - final desision on whether the protein is secreted

Examples

# read fasta file in AAStringSet object
aa <- readAAStringSet(system.file("extdata","sample_prot_100.fasta",
package = "SecretSanta"))
# create an instance of CBSResult class
sr <- CBSResult(in_fasta = aa[1:20])
# run signalpeptide prediction and create fully populated instance of
# SignalpResult class
step1_sp2 <- signalp(sr, version = 2, organism = 'euk', run_mode = "starter")

#> Version used... signalp2

#> running signalp locally...

#> 1 sequences to be truncated

#> Ok for single processing

#> Submitted sequences... 20

#> signalp < 4, calling parser for the output...

#> Candidate sequences with signal peptides... 1

# access specific slots:
getOutfasta(step1_sp2)

#>   A AAStringSet instance of length 1
#>     width seq                                               names               
#> [1]   210 MKLSSVCTIIFGLVFIDFNNVVN...MYSSDGSVTTAAQKAAALVKANA ALI_PLTG_5

getInfasta(step1_sp2)

#>   A AAStringSet instance of length 20
#>      width seq                                              names               
#>  [1]   176 MAGLGICGEPLVSGITSDLSGMS...SVTTAAQKAAALVKNSARILRA ALI_PLTG_7
#>  [2]   456 QFLSELRKVSVSTKLDGVQRLAS...AENQSQQNSRPREGCCSSCIIF ALI_PLTG_4
#>  [3]   276 MKLPTGYGCETTNGDDAASGGNS...FTGMIEDIKFLCKVDVSNEKQL ALI_PLTG_11
#>  [4]   260 MQHDEVIWSVISKQFCSFKSKLP...KGPYVEIEYEQEQETATGELNW ALI_PLTG_10
#>  [5]   168 MAPATDVKVKLVSMDGEAFEVEA...TEEFTEEEQQRILEENKWCEDV ALI_PLTG_14
#>  ...   ... ...
#> [16]  1999 MAGLGICGEPLVSGITSDLSGMS...LVKEKMELVEKLESQVASTENT ALI_PLTG_9_truncated
#> [17]   155 MSVSSYAVSYLASMDQTSAGPGA...SVTTAAQKAAALVQGGARKLRM ALI_PLTG_8
#> [18]   166 MFLNVAMSDNMHFGRELDEEVLA...VADIPFICIVLLPMKANAATGK ALI_PLTG_18
#> [19]   166 MTALLVKSSWKQHEPMPTADVLV...SSSQEAYRRRAYYPAKKEKAKK ALI_PLTG_17
#> [20]   570 MPPKKGSSTPELLEPYERHSYAD...RFYSAIKQRLQEQLLVVMQSGY ALI_PLTG_16

getSPtibble(step1_sp2)

#> # A tibble: 1 x 9
#>      gene_id  Cmax  Cpos  Ymax  Ypos  Smax  Spos Smean     Prediction
#>       <fctr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>         <fctr>
#> 1 ALI_PLTG_5     1    25 0.687    25 0.906     3 0.637 Signal peptide

getSPversion(step1_sp2)

#> [1] 2

getMatfasta(step1_sp2)

#>   A AAStringSet instance of length 1
#>     width seq                                               names               
#> [1]   186 QCSDVHVVFARGSGEAAGLGICG...MYSSDGSVTTAAQKAAALVKANA ALI_PLTG_5