accesor functions for objects of TMhmmResult S4 class, intermediate and final outputs of the tmhmm prediction step

accesor functions for objects of TMhmmResult S4 class, intermediate and final outputs of the tmhmm prediction step

Accessors for TMhmmResul objects

setTMtibble(theObject, tm_tibble)

# S4 method for TMhmmResult
setTMtibble(theObject, tm_tibble)

# S4 method for TMhmmResult
getTMtibble(theObject)

# S4 method for TMhmmResult
getInMatfasta(theObject)

# S4 method for TMhmmResult
getOutMatfasta(theObject)

Arguments

theObject
object of TMhmmREsult class

tm_tibble
parsed tmhmm output in tabular format

Value

TMhmmREsult object

Slots

in_mature_fasta

input mature fasta, extracted from the input SignalpResult object

out_mature_fasta

output mature, conatins mature sequences without TM domains

tm_tibble

tibble with outputs obtained from TMHMM

gene_id - unique id of the sequence

length - length of the protein sequence

ExpAA - the expected number of amino acids intransmembrane helices

First60 - the expected number of amino acids in transmembrane helices in the first 60 amino acids of the protein

PredHel - the number of predicted transmembrane helices by N-best

Topology - the topology predicted by N-best'

Examples

aa <- readAAStringSet(system.file("extdata", "sample_prot_100.fasta",
package = "SecretSanta"))
inp <- CBSResult(in_fasta = aa[1:10])
s1_sp2 <- signalp(inp, version = 2, organism = 'euk',
run_mode = "starter")

#> Version used... signalp2

#> running signalp locally...

#> 1 sequences to be truncated

#> Ok for single processing

#> Submitted sequences... 10

#> signalp < 4, calling parser for the output...

#> Candidate sequences with signal peptides... 1

tm <- tmhmm(s1_sp2, TM = 1)

#> running TMHMM locally...

#> Submitted sequences... 1

#> Candidates with signal peptides and 0 TM domains in mature seq 1

class(tm)

#> [1] "TMhmmResult"
#> attr(,"package")
#> [1] "SecretSanta"

getTMtibble(tm)

#> # A tibble: 1 x 6
#>      gene_id length ExpAA First60 PredHel Topology
#>        <chr>  <dbl> <dbl>   <dbl>   <dbl>    <chr>
#> 1 ALI_PLTG_5    186  1.29    1.16       0        o