accesor functions for objects of TargetpResult S4 class, intermediate and final outputs of the targetp prediction step

accesor functions for objects of TargetpResult S4 class, intermediate and final outputs of the targetp prediction step

Accessors for TargetpResult objects

setTPtibble(theObject, tp_tibble)

# S4 method for TargetpResult
setTPtibble(theObject, tp_tibble)

# S4 method for TargetpResult
getTPtibble(theObject)

Arguments

theObject
an object of TargetpResult class

tp_tibble
parsed targetp output in tabular format

Value

TargetpResult object

Slots

tp_tibble

tibble with outputs obtained from targetp

gene_id - unique id of the sequence

length - length of the protein sequence

mTP - mitochondrial NN score

sp - signal peptide NN score

other - any onther NN score

TP_localization - Prediction of localization, based on the scores; the possible values are:

C - Chloroplast, i.e. the sequence contains cTP, a chloroplast transit peptide;

M - Mitochondrion, i.e. the sequence contains mTP, a mitochondrial targeting peptide;

S - Secretory pathway, i.e. the sequence contains SP, a signal peptide;

_ - Any other location;

"don't know" - indicates that cutoff restrictions were set (see instructions) and the winning network output score was below the requested cutoff for that category.

RC - Reliability class, from 1 to 5, where 1 indicates the strongest prediction. RC is a measure of the size of the difference ('diff') between #' the highest (winning) and the second highest output scores. There are 5 reliability classes, defined as follows:

1 - diff > 0.800;

2 - 0.800 > diff > 0.600

3 - 0.600 > diff > 0.400

4 - 0.400 > diff > 0.200

5 - 0.200 > diff

Examples

#read fasta file in AAStringSet object
aa <- readAAStringSet(system.file("extdata","sample_prot_100.fasta",
package = "SecretSanta"))
#assign this object to the input_fasta slot of SignalpResult object
inp <- CBSResult(in_fasta = aa[1:10])
tp_result <- targetp(input_obj = inp, network = 'N',
run_mode = 'starter')

#> running targetp locally...

#> Ok for single processing

#> Number of submitted sequences... 10

#> Number of candidate secreted sequences 1

class(tp_result)

#> [1] "TargetpResult"
#> attr(,"package")
#> [1] "SecretSanta"

getTPtibble(tp_result)

#> # A tibble: 1 x 7
#>      gene_id length   mTP    sp other TP_localization    RC
#>       <fctr>  <int> <dbl> <dbl> <dbl>          <fctr> <int>
#> 1 ALI_PLTG_5    210 0.027 0.944 0.096               S     1

getInfasta(tp_result)

#>   A AAStringSet instance of length 10
#>      width seq                                              names               
#>  [1]    94 MVEPSSPECVVQIIKDEKLKRAD...SMTLLVCGFIYFAVLQSWLLES ALI_PLTG_1
#>  [2]   240 MTRQGYLIVHDKRSRPTVRYLSL...ATAVTPHYVPTHFGRGNNQEAH ALI_PLTG_2
#>  [3]   533 MQPTSSSAPQSDVVVADPPSVTS...LQQQQQQEQQANSTALKNERSI ALI_PLTG_3
#>  [4]   456 QFLSELRKVSVSTKLDGVQRLAS...AENQSQQNSRPREGCCSSCIIF ALI_PLTG_4
#>  [5]   210 MKLSSVCTIIFGLVFIDFNNVVN...YSSDGSVTTAAQKAAALVKANA ALI_PLTG_5
#>  [6]   177 MPGLGICGEPLVSGITSNLGGMS...VTTAAQKAAALVQGGSARALRG ALI_PLTG_6
#>  [7]   176 MAGLGICGEPLVSGITSDLSGMS...SVTTAAQKAAALVKNSARILRA ALI_PLTG_7
#>  [8]   155 MSVSSYAVSYLASMDQTSAGPGA...SVTTAAQKAAALVQGGARKLRM ALI_PLTG_8
#>  [9]  2095 MAGLGICGEPLVSGITSDLSGMS...SAIAAKNQEEPKERERPECAQQ ALI_PLTG_9
#> [10]   260 MQHDEVIWSVISKQFCSFKSKLP...KGPYVEIEYEQEQETATGELNW ALI_PLTG_10

getOutfasta(tp_result)

#>   A AAStringSet instance of length 1
#>     width seq                                               names               
#> [1]   210 MKLSSVCTIIFGLVFIDFNNVVN...MYSSDGSVTTAAQKAAALVKANA ALI_PLTG_5