organise and test external dependencies

This function checks that all the external dependencies required for secertome prediction pipelines are available and can sucessfully pass test runs. All the external dependencies can be made accessible via $PATH environment variable; alternatively a text file the with full paths to all the external dependencies should be provided. For the instalation instructions please see SecretSanta vignette.

manage_paths(in_path = c(TRUE, FALSE), test_mode = c("all", "signalp2",
  "signalp3", "signalp4", "targetp", "tmhmm", "wolfpsort"), path_file = NULL)

Arguments

in_path
if TRUE manage_paths will attempt to run external dependencies, assuming that respective paths are attached to the $PATH variable; if FALSE you should supply path_file.

test_mode
if all - all the external dependencies will be checked; alternatively specify a tool name to be checked.

path_file
full paths to external dependencies in a 2-column space-separated text file; for multiple versions of signalp please use 'signalpV'notation, where V is a version number; first column should contain tool name; second column should contain full path to the tool's executable.

Value

a list of length 3 with the following elements:

tests TRUE if all the exteranl dependencies are working;

in_path TRUE if the dependecies are accessible via $PATH; FALSE - if paths are provided with the path_file argument;

path_tibble a tibble with verified names and paths for external dependencies; NA if in_path is TRUE; if a specific tool is tested, a singualr path will be returned.

Examples

# Example1:
# here we assume that paths to all the
# external dependencies are attached to
# the $PATH variable:

manage_paths(in_path = TRUE, test_mode = 'all')

#> checking dependencies accessible via $PATH

#> signalp2 run completed

#> signalp3 run completed

#> signalp4 run completed

#> tmhmm run completed

#> targetp run completed

#> wolfpsort run completed

#> $tests
#> [1] TRUE
#> 
#> $in_path
#> [1] TRUE
#> 
#> $path_tibble
#> [1] NA
#> 

# to test just a single tool:
manage_paths(in_path = TRUE, test_mode = 'signalp2')

#> checking dependencies accessible via $PATH

#> signalp2 run completed

#> $tests
#> [1] TRUE
#> 
#> $in_path
#> [1] TRUE
#> 
#> $path_tibble
#> [1] NA
#> 

# Example2:
# alternatively, we are in a situation
# when changing $PATH is not possible,
# so we supply a file with listed full
# paths to the external dependencies:

manage_paths(in_path = FALSE,
test_mode = 'all',
path_file = system.file("extdata", "sample_paths", package = "SecretSanta"))

#> All paths are valid

#> signalp2 run completed

#> signalp3 run completed

#> signalp4 run completed

#> tmhmm run completed

#> targetp run completed

#> wolfpsort run completed

#> $tests
#> [1] TRUE
#> 
#> $in_path
#> [1] FALSE
#> 
#> $path_tibble
#> # A tibble: 6 x 3
#>        tool
#>       <chr>
#> 1  signalp4
#> 2  signalp3
#> 3  signalp2
#> 4   targetp
#> 5     tmhmm
#> 6 wolfpsort
#> # ... with 2 more variables: path <chr>, status <lgl>
#>