predict subcellular protein localization with TargetP

This function calls local targetp to predict subcellular localisation of a protein. For large inputs (>1000 sequences) will automatically run as a massive paralle job.

targetp(input_obj, network = c("P", "N"), run_mode = c("starter", "piper"),
  paths = NULL, cores = NULL)

Arguments

input_obj
an instance of CBSResult class containing protein sequences in one of the slots

network
P - for plants; N - for non-plants;

run_mode
starter - if it is the first step in pipeline; piper - if you run this function on the output of other CBS tools;

paths
if targetp is not acessible globally, a file conatining a full path to it's executable should be provided; for details please check SecretSanta vignette.

cores
optional arguments, number of cores to run the parallel process on. If not set default will be 1.

Value

an object of TargetpResult class

Examples

# read fasta file in AAStringSet object
aa <- readAAStringSet(system.file("extdata", "sample_prot_100.fasta",
package = "SecretSanta"))
# assign this object to the input_fasta 
# slot of CBSResult object
inp <- CBSResult(in_fasta = aa[1:10])
# run target prediction
tp_result <- targetp(input_obj = inp, network = 'N', run_mode = 'starter')

#> running targetp locally...

#> Ok for single processing

#> Number of submitted sequences... 10

#> Number of candidate secreted sequences 1