run_WolfPsort function

This function runs WoLF PSORT to predict protein cellular sub-localisation and returns the most probbale one. Including this step in secretome prediction pipelines provides additional supportig evidence that a protein might be secreted and deposited outside the cell. Recommended to run on the late stages of secretome prediction pipeline. Also see targetp function - for similar functionality.

wolfpsort(input_obj, organism = c("plant", "animal", "fungi"), paths = NULL)

Arguments

input_obj
Object of CSBResult class

organism
set relevant taxonomic group, options include: plant, animal, fungi;

paths
if wolfpsort is not acessible globally, a file conatining a full path to it's executable should be provided; for details please check SecretSanta vignette.

Value

object of WolfResult class

Examples

# read fasta file in AAStringSet object
aa <- readAAStringSet(system.file("extdata","sample_prot_100.fasta",
package = "SecretSanta"))
# assign this object to the input_fasta slot of CBSResult object
inp <- CBSResult(in_fasta = aa[1:10])
# run signalp2 on the initial file:
step1_sp2 <- signalp(inp, version = 2, organism ='euk',
run_mode = "starter")

#> Version used... signalp2

#> running signalp locally...

#> 1 sequences to be truncated

#> Ok for single processing

#> Submitted sequences... 10

#> signalp < 4, calling parser for the output...

#> Candidate sequences with signal peptides... 1

# run wolfpsort on the signalp output:
w <- wolfpsort(step1_sp2, 'fungi')

#> running WoLF PSORT locally...

#> Number of submitted sequences... 1

#> Candidate sequences with extracellular localisation... 1